% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sFFLHDmm.R
\docType{data}
\name{sFFLHDmm}
\alias{sFFLHDmm}
\title{sFFLHD maximin}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
sFFLHDmm
}
\value{
A sFFLHDmm object
}
\description{
sFFLHD R6 object that gives a batch of points at a time using maximin.
To do this it takes all batches for stage at beginning of stage
and then reorders them.
Not that great in practice.
Requires extra optimization and storage.
}
\section{Fields}{

\describe{
\item{\code{D}}{numeric. The number of dimensions for the design. Must be set.}

\item{\code{L}}{numeric. The number of points in each batch, also the number of
levels of each dimension. Must be set.}

\item{\code{b}}{integer. The batch number.}

\item{\code{s}}{sFFLHD. The design it takes the points and then reorders them.}

\item{\code{X}}{matrix. The points given in the design.}

\item{\code{Xchoices}}{list. Batches taken from s and have been reordered,
but which have not been returned to the user yet.}
}}

\examples{
s <- sFFLHDmm$new(D=2,L=3)
s$get.batch()
s <- sFFLHDmm$new(D=2,L=4)
s$get.batch()
}
\keyword{datasets}
